Remove sorting facilities
The sorting facilities (provided publically by stringv_sort, and
internally by quicksort and insertion sort) were clunky, and due to the
inplace requirement, inefficient. Sorting is important, as it enables
things like binary search, lower/upper/equal bounds and merge
operations.

The intention is to rewrite quicksort using a buffered approach. That
is, the public sort function takes a user buffer that it can use for
swap space and to buffer sorting.